+++ packages/design-tokens/src/index.ts
@@ -24,8 +24,6 @@
 
  */
 
-import * as palette from './tokens/color/palette'
-
 export * from './system'
 export * from './theme'
 export * from './GlobalStyle'
@@ -35,8 +33,7 @@ export {
   defaultFonts,
 } from './tokens/typography/font_families'
 
-export { palette }
-export { prismTheme } from './prismTheme'
+export { palette } from './legacy'
 
 // Useful external utilities
 export * from './utils/animations'